requirements.txt
# ExamPulse Backend Dependencies

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI and ML
google-generativeai==0.3.2

# OCR
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Database
supabase==2.0.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0

